+++ src/typespec/state.ts
@@ -14,6 +14,7 @@ export const COCOGEN_STATE_PROPERTY_TYPESPEC_DESCRIPTIONS = Symbol.for("@wictorw
 export const COCOGEN_STATE_PROPERTY_SEARCH = Symbol.for("@wictorwilen/cocogen/propertySearch");
 export const COCOGEN_STATE_PROPERTY_SOURCE = Symbol.for("@wictorwilen/cocogen/propertySource");
 export const COCOGEN_STATE_PROPERTY_SOURCE_ENTRIES = Symbol.for("@wictorwilen/cocogen/propertySourceEntries");
+export const COCOGEN_STATE_PROPERTY_SOURCE_DEFAULT = Symbol.for("@wictorwilen/cocogen/propertySourceDefault");
 export const COCOGEN_STATE_PROPERTY_NO_SOURCE = Symbol.for("@wictorwilen/cocogen/propertyNoSource");
 export const COCOGEN_STATE_PROPERTY_PERSON_FIELDS = Symbol.for("@wictorwilen/cocogen/propertyPersonFields");
 export const COCOGEN_STATE_PROPERTY_SERIALIZED = Symbol.for("@wictorwilen/cocogen/propertySerialized");
@@ -64,11 +65,13 @@ export type CocogenSearchFlags = {
 export type CocogenSourceSettings = {
   csv?: string;
   jsonPath?: string;
+  default?: string;
 };
 
 export type CocogenSourceEntry = {
   from: CocogenSourceSettings | string;
   to?: string;
+  default?: string;
 };
 
 export type CocogenSerializedTarget = {
@@ -78,5 +81,6 @@ export type CocogenSerializedTarget = {
 export type CocogenPersonEntityField = {
   path: string;
   source: CocogenSourceSettings | string;
+  default?: string;
 };