+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support `default` values in `@coco.source` mappings via the second argument object.
+
 ## [1.0.36] - 2026-02-03
 
 ### Added